+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 * Temporarily disabled `tag` in target selectors due to a strange Minecraft edge case.
-* 
+* Fixed an issue with propagating macros.
+* Fixed #2.